Releasing using OpenJDK (latest) instead of 8uX.
The infra structure used to publish releases (EF Jenkins) from time to
time changes its set of supported JDK. This includes removal of some
platforms, like Open JDK 8. So there is an ongoing need to update the
Jenkinsfile, which is a burden without actual benefit.

Instead of this ongoing changes, we will use the latest Open JDK release
always, which is done by using the "latest" label of the JDK
declaration.