Deprecating "APPLICATION/SVG+XML" for Removal
"APPLICATION/SVG+XML" is not recognized by modern browsers, as the official MIME type of SVG graphics is "IMAGE/SVG+XML".

As least RESTful applications produce SVG at all, it makes not much sense to keep this and add the new variant.
Instead, a custom String or a custom library should replace the outdated elements.